Update clientconnection.py
Update for the new error messages issued by what.cd: They added the "reason" for the report. Because the reason is always added at the end, we can use re.match instead of re.search, which is more efficient.

Might be better to catch the reason for the report and show it to the user, but I'm not qualified to do so.

Anyway, we fixed the error. So uhm... Peace !